prompt-engineering/api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
mistralai==1.6.0
openai==1.70.0
requests==2.31.0
tiktoken==0.9.0
rich==14.0.0